+++ sim/abstract_controller.hh
@@ -37,8 +37,8 @@ class AbstractController : public Object {
 
   ControllerID getControllerID() { return controllerID; }
 
-  virtual uint64_t read(uint64_t, uint64_t, uint8_t *) noexcept = 0;
-  virtual uint64_t write(uint64_t, uint64_t, uint8_t *) noexcept = 0;
+  virtual uint64_t read(uint64_t, uint32_t, uint8_t *) noexcept = 0;
+  virtual uint64_t write(uint64_t, uint32_t, uint8_t *) noexcept = 0;
 
   void createCheckpoint(std::ostream &out) const noexcept override {
     BACKUP_SCALAR(out, controllerID);
